Change Decimal field To Double

+++ Shoping/Business/ProductServices/IProductBusiness.cs
@@ -13,7 +13,7 @@ namespace Shoping.Business.ProductServices
         public Task<Guid> AddUpdateProductAsync(ProductDTO productDTO);
         public Task<bool> DeleteProductAsync(Guid productRecID);
         public Task<PageData<ProductDTO>> GetProductsPaging(int page, int pageSize);
-        public Task<PageData<ProductDTO>> GetFilterProducts(String search, Guid CatID, decimal from, decimal to, int page, int pageSize);
+        public Task<PageData<ProductDTO>> GetFilterProducts(String search, Guid CatID, double from, double to, int page, int pageSize);
         public Task<List<ProductDTO>> GetListProductsByRecID(List<Guid> lstRecIDs);
         public Task<bool> DeleteAllProducts();
     }
