+++ Shoping/Data Access/DTOs/ProductDTO.cs
@@ -12,10 +12,10 @@ namespace Shoping.Data_Access.DTOs
         public Guid RecID { get; set; }
         public Guid ProductID { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string Image { get; set; } // Implement later
         public Guid CatID { get; set; } 
-        public decimal PurchasePrice { get; set; } 
+        public double PurchasePrice { get; set; } 
         public int Quantity { get; set; } 
         public object Clone()
         {
